io module shuffling, move common infrastructure to MyTake
